Redis/requirements.txt
redis==2.10.5
kafka-python==1.3.1
futures==3.0.5
six==1.10.0
wheel==0.24.0